Feat: implement user registration and password hashing

+++ main.py
@@ -11,7 +11,7 @@ from fastapi.staticfiles import StaticFiles
 
 from core.SendFile import Send, Split
 from core.GetFile import Get, Merge
-from database.Database import create_db_and_tables, add_user_to_db, add_file_to_db, get_user_files, get_file_by_id, delete_file
+from database.Database import create_db_and_tables, add_user_to_db, add_file_to_db, get_user_files, get_file_by_id, delete_file, get_user_by_name, verify_password
 
 load_dotenv()
 
@@ -35,52 +35,94 @@ app.mount("/assets", StaticFiles(directory="assets"), name="assets")
 templates = Jinja2Templates(directory="templates")
 
 @app.get("/", response_class=HTMLResponse)
-async def home(request: Request):
-    return templates.TemplateResponse(request=request, name="index.html") # load the HTML
+async def login_page(request: Request):
+    if request.cookies.get("session_username"):
+        return RedirectResponse(url="/dashboard", status_code=303)
+    return templates.TemplateResponse(request=request, name="login.html", context={})
 
 
 @app.post("/register")
-async def register(username: str = Form(...)):
-    await add_user_to_db(username)
-    return RedirectResponse(url=f"/dashboard?username={username}", status_code=303) # load the user and return a 303 status
+async def register(request: Request, username: str = Form(...), password: str = Form(...)):
+    existing_user = await get_user_by_name(username)
+    if existing_user:
+        return templates.TemplateResponse(
+            request=request, 
+            name="login.html", 
+            context={"error": "This username is already taken"}
+        )
 
+    try:
+        await add_user_to_db(username, password)
+    except Exception as e:
+        print(f"Ошибка регистрации: {e}")
+        return templates.TemplateResponse(
+            request=request, 
+            name="login.html", 
+            context={"error": "Error creating account"}
+        )
+    # Autologin (If the user is in the cookie files)
+    response = RedirectResponse(url="/dashboard", status_code=303)
+    response.set_cookie(key="session_username", value=username, httponly=True, max_age=86400)
+    return response
 
 @app.post("/login")
-async def login(username: str = Form(...)):
-    return RedirectResponse(url=f"/dashboard?username={username}", status_code=303) # load the user and return a 303 status
-
+async def login_post(request: Request, username: str = Form(...), password: str = Form(...)):
+    user = await get_user_by_name(username)
+    
+    if user and verify_password(password, user.password_hash):
+        response = RedirectResponse(url="/dashboard", status_code=303)
+        response.set_cookie(key="session_username", value=username, httponly=True, max_age=86400)
+        return response
+    else:
+        return templates.TemplateResponse(
+            request=request, 
+            name="login.html", 
+            context={"error": "Incorrect username or password"}
+        )
 
 @app.get("/dashboard", response_class=HTMLResponse)
-async def dashboard(request: Request, username: str):
-    files_list = await get_user_files(username)
-
+async def dashboard(request: Request):
+    username = request.cookies.get("session_username")
+    if not username:
+        return RedirectResponse(url="/", status_code=303)
+    
+    files = await get_user_files(username)
+    
     return templates.TemplateResponse(
         request=request, 
         name="dashboard.html", 
-        context={
-            "username": username, 
-            "files": files_list
-        }
+        context={"username": username, "files": files}
     )
+
+@app.get("/logout")
+async def logout():
+    response = RedirectResponse(url="/", status_code=303)
+    # delete cookies
+    response.delete_cookie("session_username")
+    return response
+
 @app.post("/upload")
-async def upload_file(username: str = Form(...), file: UploadFile = File(...)):
+async def upload_file(request: Request, file: UploadFile = File(...)):
+    username = request.cookies.get("session_username")
+    if not username:
+        raise HTTPException(status_code=401)
+
     temp_path = f"temp_{file.filename}"
     with open(temp_path, "wb") as buffer:
         buffer.write(await file.read())
 
-    # split file
     chunks = await Split(temp_path)
     currentid_chunk_list = []
     
     for chunk in chunks: 
         message = await Send(client, "me", chunk)
         currentid_chunk_list.append(message.id)
-        os.remove(chunk) # delete chunk
+        os.remove(chunk)
 
-    await add_file_to_db(username, file.filename, currentid_chunk_list) # added to datebase
-    os.remove(temp_path) # delete original file
-    # Back to the dashboard
-    return RedirectResponse(url=f"/dashboard?username={username}", status_code=303)
+    await add_file_to_db(username, file.filename, currentid_chunk_list)
+    os.remove(temp_path)
+    
+    return RedirectResponse(url="/dashboard", status_code=303)
 
 
 @app.post("/download")
